requirements.txt
Pillow
pyairtable
polars
amulet-core
beautifulsoup4 
lxml
flask
flask-wtf
flask-uploads
python-dotenv
numpy
tqdm